backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.3.0
anthropic==0.7.8
aiohttp==3.9.0
requests==2.31.0
click==8.1.7